Added new helper function and tidy up
